Added Upload That Thang
This is a small self hosted uploading service i used, because it was easier than using something like guacamole. I will add folder support to the uploader and maybe also multithreading

Every uploaded file will be located in the home directory. This is currently hard-coded but i will add support to change the upload directory.

The name comes from a project i was working on but scraped because i remembered i could just intigrate it here.